+++ src/main/java/com/chordplay/chordplayapiserver/domain/sheet/listener/SheetRedisListener.java
@@ -1,4 +1,4 @@
-package com.chordplay.chordplayapiserver.domain.sheet;
+package com.chordplay.chordplayapiserver.domain.sheet.listener;
 
 import com.chordplay.chordplayapiserver.domain.sheet.dto.AiStatusMessage;
 import com.chordplay.chordplayapiserver.global.sse.CustomSseEmitter;
